feat: next config template
