+++ templates/base/packages/nextjs/next.config.js.template.mjs
@@ -1,0 +1,29 @@
+import { withDefaults } from '../../../utils.js'
+
+const contents = ({ ignoreTsAndLintBuildErrors, extraConfig }) =>
+`// @ts-check
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  ${extraConfig[0] ? `${extraConfig[0]},` : ''}
+  reactStrictMode: true,
+  typescript: {
+    ignoreBuildErrors: ${ignoreTsAndLintBuildErrors},
+  },
+  eslint: {
+    ignoreDuringBuilds: ${ignoreTsAndLintBuildErrors},
+  },
+  webpack: config => {
+    config.resolve.fallback = { fs: false, net: false, tls: false };
+    config.externals.push("pino-pretty", "lokijs", "encoding");
+    return config;
+  },
+};
+
+module.exports = nextConfig;
+`
+
+export default withDefaults(contents, {
+  ignoreTsAndLintBuildErrors: 'process.env.NEXT_PUBLIC_IGNORE_BUILD_ERROR === "true"',
+  extraConfig: ''
+})
